fix: add snippets in note scripts to ensure account storage changes
This is a quick fix that ensures the notes used in `unask_community_cards` always end up modifying the account's storage. This is a requirement because the current tx script that consumes notes increments the nonce and when doing so either the account storage should change or the account's vault should change.