wip: add custom tx for consuming notes ensuring account storage changes
This is a fix that ensures that consuming notes always end up modifying the account's storage, even if the note does not modify it. This is a requirement because the current tx script that consumes notes increments the nonce and when doing so either the account storage should change or the account's vault should change.